backend/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg==3.2.6
psycopg2-binary==2.9.10
python-dotenv==1.1.0
pytz==2025.2
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==3.1.3
